refactor(duke-gc): Extract HostManager out of Heap God Struct
Extracted OS-level file, process, and socket management logic out of `crates/duke-gc/src/lib.rs` into a new `HostManager` struct in `crates/duke-gc/src/host.rs` to enforce clear domain boundaries.
- Created `host.rs` module.
- Moved `HostFileHandle`, `HostProcessHandle`, and `SpawnedProcessIds` to `host.rs`.
- Added `host: HostManager` to `Heap`.
- Updated all call-sites in `duke-interpreter` to use `heap.host.<method>()`.
