+++ crates/duke-interpreter/src/lib.rs
@@ -2533,7 +2533,7 @@ fn native_file_input_stream_init(
         _ => return Err(VmError::NullPointerException),
     };
     let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_input_file(&path)?;
+    let file_id = heap.host.open_host_input_file(&path)?;
     let stream_obj = heap.get_mut(this_ref)?;
     let Some(fd_field) = stream_obj.fields.first_mut() else {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2549,7 +2549,7 @@ fn native_file_input_stream_read(
     _control: &mut NativeControl,
 ) -> VmResult<Option<Slot>> {
     let file_id = file_stream_id_from_this(args, heap)?;
-    Ok(Some(Slot::Int(heap.read_host_file_byte(file_id)?)))
+    Ok(Some(Slot::Int(heap.host.read_host_file_byte(file_id)?)))
 }
 
 fn native_file_input_stream_read_bytes(
@@ -2570,7 +2570,7 @@ fn native_file_input_stream_read_bytes(
 
     let mut count = 0_usize;
     for idx in 0..len {
-        let next = heap.read_host_file_byte(file_id)?;
+        let next = heap.host.read_host_file_byte(file_id)?;
         if next < 0 {
             break;
         }
@@ -2603,7 +2603,7 @@ fn native_file_input_stream_close(
             });
         }
     };
-    heap.close_host_file(file_id);
+    heap.host.close_host_file(file_id);
     let stream_obj = heap.get_mut(this_ref)?;
     let Some(fd_field) = stream_obj.fields.first_mut() else {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2623,7 +2623,7 @@ fn native_file_output_stream_init(
         _ => return Err(VmError::NullPointerException),
     };
     let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_output_file(&path)?;
+    let file_id = heap.host.open_host_output_file(&path)?;
     let stream_obj = heap.get_mut(this_ref)?;
     let Some(fd_field) = stream_obj.fields.first_mut() else {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2648,7 +2648,7 @@ fn native_file_output_stream_write(
             });
         }
     };
-    heap.write_host_file_byte(file_id, value)?;
+    heap.host.write_host_file_byte(file_id, value)?;
     Ok(None)
 }
 
@@ -2671,7 +2671,7 @@ fn native_file_output_stream_write_bytes(
                 got: "other",
             });
         };
-        heap.write_host_file_byte(file_id, value)?;
+        heap.host.write_host_file_byte(file_id, value)?;
     }
     Ok(None)
 }
@@ -2694,7 +2694,7 @@ fn native_file_output_stream_close(
             });
         }
     };
-    heap.close_host_file(file_id);
+    heap.host.close_host_file(file_id);
     let stream_obj = heap.get_mut(this_ref)?;
     let Some(fd_field) = stream_obj.fields.first_mut() else {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2726,8 +2726,8 @@ fn native_server_socket_init(
         }
     };
     let addr = format!("0.0.0.0:{port}");
-    let server_id = heap.bind_server_socket(&addr)?;
-    let actual_port = heap.server_socket_local_port(server_id)?;
+    let server_id = heap.host.bind_server_socket(&addr)?;
+    let actual_port = heap.host.server_socket_local_port(server_id)?;
     let obj = heap.get_mut(this_ref)?;
     if obj.fields.len() < 2 {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2756,7 +2756,7 @@ fn native_server_socket_accept(
             });
         }
     };
-    let (reader_id, writer_id) = heap.accept_connection(server_fd)?;
+    let (reader_id, writer_id) = heap.host.accept_connection(server_fd)?;
     // Allocate a new Socket object with fdRead=reader_id, fdWrite=writer_id
     let socket_ref = heap.allocate("java/net/Socket".to_string(), 2);
     heap.get_mut(socket_ref)?.fields[0] = Slot::Int(reader_id);
@@ -2803,7 +2803,7 @@ fn native_server_socket_close(
             });
         }
     };
-    heap.close_host_file(fd);
+    heap.host.close_host_file(fd);
     let obj = heap.get_mut(this_ref)?;
     obj.fields[0] = Slot::Int(0);
     obj.fields[1] = Slot::Int(0); // also zero cached port so getLocalPort() returns 0 after close
@@ -2839,7 +2839,7 @@ fn native_socket_init(
         }
     };
     let addr = format!("{host}:{port}");
-    let (reader_id, writer_id) = heap.connect_socket(&addr)?;
+    let (reader_id, writer_id) = heap.host.connect_socket(&addr)?;
     let obj = heap.get_mut(this_ref)?;
     if obj.fields.len() < 2 {
         return Err(VmError::InvalidRef { address: this_ref });
@@ -2924,8 +2924,8 @@ fn native_socket_close(
             });
         }
     };
-    heap.close_host_file(fd_read);
-    heap.close_host_file(fd_write);
+    heap.host.close_host_file(fd_read);
+    heap.host.close_host_file(fd_write);
     let obj = heap.get_mut(this_ref)?;
     obj.fields[0] = Slot::Int(0);
     obj.fields[1] = Slot::Int(0);
@@ -2955,7 +2955,7 @@ fn native_zip_file_init(
         .as_deref()
         .ok_or(VmError::NullPointerException)?
         .to_string();
-    let fd = heap.open_host_zip(std::path::Path::new(&path_str))?;
+    let fd = heap.host.open_host_zip(std::path::Path::new(&path_str))?;
     let obj = heap.get_mut(this_ref)?;
     obj.fields[0] = Slot::Int(fd);
     Ok(None)
@@ -2991,7 +2991,7 @@ fn native_zip_file_get_entry(
         .as_deref()
         .ok_or(VmError::NullPointerException)?
         .to_string();
-    let info = heap.zip_get_entry_info(fd, &entry_name)?;
+    let info = heap.host.zip_get_entry_info(fd, &entry_name)?;
     let Some(info) = info else {
         return Ok(Some(Slot::Reference(None)));
     };
@@ -3043,8 +3043,8 @@ fn native_zip_file_get_input_stream(
         .ok_or(VmError::NullPointerException)?
         .to_string();
     // Decompress the entry and wrap in a ByteBuffer.
-    let data = heap.zip_read_entry(fd, &entry_name)?;
-    let buf_fd = heap.open_host_byte_buffer(data);
+    let data = heap.host.zip_read_entry(fd, &entry_name)?;
+    let buf_fd = heap.host.open_host_byte_buffer(data);
     let is_ref = heap.allocate("duke/zip/ByteBufferInputStream".to_string(), 1);
     let is_obj = heap.get_mut(is_ref)?;
     is_obj.fields[0] = Slot::Int(buf_fd);
@@ -3066,7 +3066,7 @@ fn native_zip_file_close(
         Some(Slot::Int(id)) => *id,
         _ => return Ok(None),
     };
-    heap.close_host_file(fd);
+    heap.host.close_host_file(fd);
     let obj = heap.get_mut(this_ref)?;
     obj.fields[0] = Slot::Int(0);
     Ok(None)
@@ -3092,7 +3092,7 @@ fn native_zip_file_size(
             });
         }
     };
-    let count = heap.zip_entry_count(fd)?;
+    let count = heap.host.zip_entry_count(fd)?;
     Ok(Some(Slot::Int(count as i32)))
 }
 
@@ -12919,7 +12919,7 @@ fn optional_file_path_from_slot(
 
 fn allocate_process_impl(
     heap: &mut duke_gc::Heap,
-    ids: duke_gc::SpawnedProcessIds,
+    ids: duke_gc::host::SpawnedProcessIds,
 ) -> VmResult<Option<Slot>> {
     let process_ref = heap.allocate("java/lang/ProcessImpl".to_string(), 4);
     let process_obj = heap.get_mut(process_ref)?;
@@ -12935,7 +12935,7 @@ fn spawn_process_impl(
     command: &[String],
     cwd: Option<&std::path::Path>,
 ) -> VmResult<Option<Slot>> {
-    let ids = heap.spawn_host_process(command, cwd)?;
+    let ids = heap.host.spawn_host_process(command, cwd)?;
     allocate_process_impl(heap, ids)
 }
 
@@ -13111,7 +13111,7 @@ fn native_process_wait_for(
     _control: &mut NativeControl,
 ) -> VmResult<Option<Slot>> {
     let process_id = process_field_id_from_this(args, heap, PROCESS_ID_FIELD)?;
-    Ok(Some(Slot::Int(heap.wait_host_process(process_id)?)))
+    Ok(Some(Slot::Int(heap.host.wait_host_process(process_id)?)))
 }
 
 fn native_process_exit_value(
@@ -13121,7 +13121,7 @@ fn native_process_exit_value(
     _control: &mut NativeControl,
 ) -> VmResult<Option<Slot>> {
     let process_id = process_field_id_from_this(args, heap, PROCESS_ID_FIELD)?;
-    let Some(exit_code) = heap.try_host_process_exit_value(process_id)? else {
+    let Some(exit_code) = heap.host.try_host_process_exit_value(process_id)? else {
         return Err(VmError::JavaException {
             class_name: "java/lang/IllegalThreadStateException".into(),
         });
@@ -13136,7 +13136,7 @@ fn native_process_destroy(
     _control: &mut NativeControl,
 ) -> VmResult<Option<Slot>> {
     let process_id = process_field_id_from_this(args, heap, PROCESS_ID_FIELD)?;
-    heap.destroy_host_process(process_id)?;
+    heap.host.destroy_host_process(process_id)?;
     Ok(None)
 }
